Add forms model validation to import
