+++ home/models.py
@@ -1253,7 +1253,7 @@ class AnswerBlock(blocks.StructBlock):
     score = blocks.FloatBlock(
         help_text="How much to add to the total score if this answer is chosen"
     )
-    semantic_id = blocks.TextBlock(help_text="Semantic ID for this answer")
+    answer_semantic_id = blocks.TextBlock(help_text="Semantic ID for this answer")
 
 
 class BaseQuestionBlock(blocks.StructBlock):
@@ -1266,7 +1266,7 @@ class BaseQuestionBlock(blocks.StructBlock):
         required=False,
         help_text="Error message for this question if we don't understand the input",
     )
-    semantic_id = blocks.TextBlock(help_text="Semantic ID for this question")
+    question_semantic_id = blocks.TextBlock(help_text="Semantic ID for this question")
 
 
 class CategoricalQuestionBlock(BaseQuestionBlock):
