+++ home/tests/import-export-data/assessment_simple.csv
@@ -1,4 +1,4 @@
 title,question_type,tags,slug,version,locale,high_result_page,high_inflection,medium_result_page,medium_inflection,low_result_page,skip_threshold,skip_high_result_page,generic_error,question,explainer,error,min,max,answers,scores,answer_semantic_ids,question_semantic_id
-Health Assessment Simple,categorical_question,checker,Health-assessment-simple,v1.0,en,high-inflection,4.0,medium-score,2.0,low-score,2.0,skip-score,"Sorry, we didn't quite get that.",How often do you check your blood pressure,We need to know your blood pressure for a proper risk assessment,Sorry we don't under stand.,,,"Once a week,Twice a week,Thrice a week","1.0,3.5,4.5","test-1,test-2,test-3",question_semantic_id
-Health Assessment Simple,age_question,checker,Health-assessment-simple,v1.0,en,high-inflection,4.0,medium-score,2.0,low-score,2.0,skip-score,"Sorry, we didn't quite get that.",How often do you check your blood pressure,We need to know your blood pressure for a proper risk assessment,Sorry we don't under stand.,,,,,,
-Health Assessment Simple,multiselect_question,checker,Health-assessment-simple,v1.0,en,high-inflection,4.0,medium-score,2.0,low-score,2.0,skip-score,"Sorry, we didn't quite get that.",How often do you check your blood pressure,We need to know your blood pressure for a proper risk assessment,Sorry we don't under stand.,,,"Once a week,Twice a week,Thrice a week","1.0,3.5,4.5","test-1,test-2,test-3",question_semantic_id
+Health Assessment Simple,categorical_question,checker,Health-assessment-simple,v1.0,en,,4.0,,2.0,,2.0,,"Sorry, we didn't quite get that.",How often do you check your blood pressure,We need to know your blood pressure for a proper risk assessment,Sorry we don't under stand.,,,"Once a week,Twice a week,Thrice a week","1.0,3.5,4.5","test-1,test-2,test-3",question_semantic_id
+Health Assessment Simple,age_question,checker,Health-assessment-simple,v1.0,en,,4.0,,2.0,,2.0,,"Sorry, we didn't quite get that.",How often do you check your blood pressure,We need to know your blood pressure for a proper risk assessment,Sorry we don't under stand.,,,,,,question_semantic_id
+Health Assessment Simple,multiselect_question,checker,Health-assessment-simple,v1.0,en,,4.0,,2.0,,2.0,,"Sorry, we didn't quite get that.",How often do you check your blood pressure,We need to know your blood pressure for a proper risk assessment,Sorry we don't under stand.,,,"Once a week,Twice a week,Thrice a week","1.0,3.5,4.5","test-1,test-2,test-3",question_semantic_id
