+++ home/tests/import-export-data/bulk_assessments.csv
@@ -22,5 +22,5 @@ Random Choice,categorical_question,"Random,selection",random-choice,v1.0,en,high
 Health Assessment,categorical_question,checker,health-assessment,v1.0,en,high-inflection,4.0,medium-score,2.0,low-score,2.0,skip-score,"Sorry, we didn't quite get that.",How often do you check your blood pressure,We need to know this for a proper risk assessment,Sorry we don't understand.,,,"Once a week,Twice a week,Thrice a week","1.0,3.0,4.5","test-1,test-2,test-3",question_semantic_id
 Health Assessment,categorical_question,checker,health-assessment,v1.0,en,high-inflection,4.0,medium-score,2.0,low-score,2.0,skip-score,"Sorry, we didn't quite get that.",Have you been vaccinated against COVID-19?,We need to know this for a proper risk assessment,Sorry we didn't quite catch that.,,,"Yes,No","3.0,1.0","test-1,test-2",question_semantic_id
 Health Assessment,categorical_question,checker,health-assessment,v1.0,en,high-inflection,4.0,medium-score,2.0,low-score,2.0,skip-score,"Sorry, we didn't quite get that.",How high is your temperature?,We need to know this for a proper risk assessment,Please choose the option that matches your answer.,,,"Between 37.5C and 38C,Between 38.1C and 40C,Between 40.1C and 41.1C","3.0,2.0,1.0","test-1,test-2,test-3",question_semantic_id
-Integer Type Question,integer_question,integer-type-question,integer-type-question,v1.0,en,high-inflection,5.0,medium-score,3.0,low-score,2.0,skip-score,This is a generic error,How much do you weigh in kilograms?,We need to test integer type questions,Sorry your min and max weight should be between 40 and 500kg.,40,500,,,,
-Weather Trivia,integer_question,weather-trivia,weather-trivia,v1.0,en,,,,,,0.0,,"Sorry, we didn't quite get that.",What's the coldest weather you're experienced?,We need to know some things,Your reply should be between {min} and {max},20,-70,,,,
+Integer Type Question,integer_question,integer-type-question,integer-type-question,v1.0,en,high-inflection,5.0,medium-score,3.0,low-score,2.0,skip-score,This is a generic error,How much do you weigh in kilograms?,We need to test integer type questions,Sorry your min and max weight should be between 40 and 500kg.,40,500,,,,question_semantic_id
+Weather Trivia,integer_question,weather-trivia,weather-trivia,v1.0,en,,,,,,0.0,,"Sorry, we didn't quite get that.",What's the coldest weather you're experienced?,We need to know some things,Your reply should be between {min} and {max},20,70,,,,question_semantic_id
