+++ home/tests/import-export-data/multiple_assessments.csv
@@ -1,11 +1,11 @@
 title,question_type,tags,slug,version,locale,high_result_page,high_inflection,medium_result_page,medium_inflection,low_result_page,skip_threshold,skip_high_result_page,generic_error,question,explainer,error,min,max,answers,scores,answer_semantic_ids,question_semantic_id
-Random Choice,categorical_question,"Random,selection",random-choice,v1.0,en,high-inflection,4.0,medium-score,3.0,low-score,2.0,skip-score,We did not quite get that,Select a number for the first question,We need to know this for a proper risk assessment,Please retry,,,"A,B","4.0,1.0","test-1,test-2",question_semantic_id
-Random Choice,categorical_question,"Random,selection",random-choice,v1.0,en,high-inflection,4.0,medium-score,3.0,low-score,2.0,skip-score,We did not quite get that,Select a number for the second question,We need to know this for a proper risk assessment,We did not quite get that,,,"A,B,C","1.0,3.0,4.0","test-1,test-2,test-3",question_semantic_id
-Random Choice,categorical_question,"Random,selection",random-choice,v1.0,en,high-inflection,4.0,medium-score,3.0,low-score,2.0,skip-score,We did not quite get that,Select a number for the third question,We need to know this for a proper risk assessment,Sorry we did not quite get that,,,"A,B,C","2.0,1.0,2.0","test-1,test-2,test-3",question_semantic_id
-Health Assessment,categorical_question,checker,health-assessment,v1.0,en,high-inflection,4.0,medium-score,2.0,low-score,2.0,skip-score,"Sorry, we didn't quite get that.",How often do you check your blood pressure,We need to know this for a proper risk assessment,Sorry we don't understand.,,,"Once a week,Twice a week,Thrice a week","1.0,3.0,4.5","test-1,test-2,test-3",question_semantic_id
-Health Assessment,categorical_question,checker,health-assessment,v1.0,en,high-inflection,4.0,medium-score,2.0,low-score,2.0,skip-score,"Sorry, we didn't quite get that.",Have you been vaccinated against COVID-19?,We need to know this for a proper risk assessment,Sorry we didn't quite catch that.,,,"Yes,No","3.0,1.0","test-1,test-2",question_semantic_id
-Health Assessment,categorical_question,checker,health-assessment,v1.0,en,high-inflection,4.0,medium-score,2.0,low-score,2.0,skip-score,"Sorry, we didn't quite get that.",How high is your temperature?,We need to know this for a proper risk assessment,Please choose the option that matches your answer.,,,"Between 37.5C and 38C,Between 38.1C and 40C,Between 40.1C and 41.1C","3.0,2.0,1.0","test-1,test-2,test-3",question_semantic_id
-Freetext Test,freetext_question,Free-text,freetext-test,v1.0,en,high-inflection,5.0,medium-score,3.0,low-score,2.0,skip-score,Generic error message,Did you find the information useful,We need to know if the information was useful,,,,,,,
-Integer Type Question,integer_question,integer-type-question,integer-type-question,v1.0,en,high-inflection,5.0,medium-score,3.0,low-score,2.0,skip-score,This is a generic error,How much do you weigh in kilograms?,We need to test integer type questions,Sorry your min and max weight should be between 40 and 500kg.,40,500,,,,
-Year of Birth,year_of_birth_question,year-of-birth,year-of-birth,v1.0,en,high-inflection,5.0,medium-score,3.0,low-score,2.0,skip-score,This is a generic error,What's your year of birth,We need to know some things,Your year of birth must be between {current_year} and {lower_bound},,,,,,
-Weather Trivia,integer_question,weather-trivia,weather-trivia,v1.0,en,,,,,,0.0,,"Sorry, we didn't quite get that.",What's the coldest weather you're experienced?,We need to know some things,Your reply should be between {min} and {max},20,-70,,,,
+Random Choice,categorical_question,"Random,selection",random-choice,v1.0,en,,4.0,,3.0,,2.0,,We did not quite get that,Select a number for the first question,We need to know this for a proper risk assessment,Please retry,,,"A,B","4.0,1.0","test-1,test-2",question_semantic_id
+Random Choice,categorical_question,"Random,selection",random-choice,v1.0,en,,4.0,,3.0,,2.0,,We did not quite get that,Select a number for the second question,We need to know this for a proper risk assessment,We did not quite get that,,,"A,B,C","1.0,3.0,4.0","test-1,test-2,test-3",question_semantic_id
+Random Choice,categorical_question,"Random,selection",random-choice,v1.0,en,,4.0,,3.0,,2.0,,We did not quite get that,Select a number for the third question,We need to know this for a proper risk assessment,Sorry we did not quite get that,,,"A,B,C","2.0,1.0,2.0","test-1,test-2,test-3",question_semantic_id
+Health Assessment,categorical_question,checker,health-assessment,v1.0,en,,4.0,,2.0,,2.0,,"Sorry, we didn't quite get that.",How often do you check your blood pressure,We need to know this for a proper risk assessment,Sorry we don't understand.,,,"Once a week,Twice a week,Thrice a week","1.0,3.0,4.5","test-1,test-2,test-3",question_semantic_id
+Health Assessment,categorical_question,checker,health-assessment,v1.0,en,,4.0,,2.0,,2.0,,"Sorry, we didn't quite get that.",Have you been vaccinated against COVID-19?,We need to know this for a proper risk assessment,Sorry we didn't quite catch that.,,,"Yes,No","3.0,1.0","test-1,test-2",question_semantic_id
+Health Assessment,categorical_question,checker,health-assessment,v1.0,en,,4.0,,2.0,,2.0,,"Sorry, we didn't quite get that.",How high is your temperature?,We need to know this for a proper risk assessment,Please choose the option that matches your answer.,,,"Between 37.5C and 38C,Between 38.1C and 40C,Between 40.1C and 41.1C","3.0,2.0,1.0","test-1,test-2,test-3",question_semantic_id
+Freetext Test,freetext_question,Free-text,freetext-test,v1.0,en,,5.0,,3.0,,2.0,,Generic error message,Did you find the information useful,We need to know if the information was useful,,,,,,,question_semantic_id
+Integer Type Question,integer_question,integer-type-question,integer-type-question,v1.0,en,,5.0,,3.0,,2.0,,This is a generic error,How much do you weigh in kilograms?,We need to test integer type questions,Sorry your min and max weight should be between 40 and 500kg.,40,500,,,,question_semantic_id
+Year of Birth,year_of_birth_question,year-of-birth,year-of-birth,v1.0,en,,5.0,,3.0,,2.0,,This is a generic error,What's your year of birth,We need to know some things,Your year of birth must be between {current_year} and {lower_bound},,,,,,question_semantic_id
+Weather Trivia,integer_question,weather-trivia,weather-trivia,v1.0,en,,,,,,0.0,,"Sorry, we didn't quite get that.",What's the coldest weather you're experienced?,We need to know some things,Your reply should be between {min} and {max},20,70,,,,question_semantic_id
