+++ home/tests/import-export-data/results_assessments.csv
@@ -23,5 +23,5 @@ Health Assessment,categorical_question,checker,health-assessment,v1.0,en,high-in
 Health Assessment,categorical_question,checker,health-assessment,v1.0,en,high-inflection,4.0,medium-score,2.0,low-score,2.0,skip-score,"Sorry, we didn't quite get that.",Have you been vaccinated against COVID-19?,We need to know this for a proper risk assessment,Sorry we didn't quite catch that.,,,"Yes,No","3.0,1.0","test-1,test-2",question_semantic_id
 Health Assessment,categorical_question,checker,health-assessment,v1.0,en,high-inflection,4.0,medium-score,2.0,low-score,2.0,skip-score,"Sorry, we didn't quite get that.",How high is your temperature?,We need to know this for a proper risk assessment,Please choose the option that matches your answer.,,,"Between 37.5C and 38C,Between 38.1C and 40C,Between 40.1C and 41.1C","3.0,2.0,1.0","test-1,test-2,test-3",question_semantic_id
 Integer Type Question,integer_question,integer-type-question,integer-type-question,v1.0,en,high-inflection,5.0,medium-score,3.0,low-score,2.0,skip-score,This is a generic error,How much do you weigh in kilograms?,We need to test integer type questions,Sorry your min and max weight should be between 40 and 500kg.,40,500,,,,question_semantic_id
-Year of Birth,year_of_birth_question,year-of-birth,year-of-birth,v1.0,en,,,,,,0.0,,This a generic error,What's your year of birth?,We need to know some things,You entered an invalid year,,,,,,
-Freetext Type Question,freetext_question,freetext-type-question,freetext-type-question,v1.0,en,,,,,,0.0,,This a generic error,Can you provide feedback on service rating?,We need to know some things,,,,,,,
+Year of Birth,year_of_birth_question,year-of-birth,year-of-birth,v1.0,en,,,,,,0.0,,This a generic error,What's your year of birth?,We need to know some things,You entered an invalid year,,,,,,question_semantic_id
+Freetext Type Question,freetext_question,freetext-type-question,freetext-type-question,v1.0,en,,,,,,0.0,,This a generic error,Can you provide feedback on service rating?,We need to know some things,,,,,,,question_semantic_id
