+++ home/tests/test_assessment_import_export.py
@@ -394,11 +394,12 @@ class TestImportExportRoundtrip:
                     "value": {
                         "question": "test question",
                         "error": "test error",
+                        "question_semantic_id": "unique_id",
                         "answers": [
                             {
                                 "answer": "test answer",
                                 "score": 2.0,
-                                "semantic_id": "unique_id",
+                                "answer_semantic_id": "unique_id",
                             }
                         ],
                     },
