+++ home/tests/test_serializers.py
@@ -145,7 +145,7 @@ def create_form_with_fields() -> Assessment:
                     "question": "test question",
                     "explainer": "We need to know this",
                     "error": "test error",
-                    "semantic_id": "test-semantic-id",
+                    "question_semantic_id": "test-semantic-id",
                     "answers": [
                         {
                             "id": "eb96ad43-c231-4493-a235-de88e60219ea",
@@ -196,7 +196,7 @@ def create_form_with_imported_fields() -> Assessment:
                     "question": "test question",
                     "explainer": "We need to know this",
                     "error": "test error",
-                    "semantic_id": "test-semantic-id",
+                    "question_semantic_id": "test-semantic-id",
                     "answers": [
                         {"score": 2.0, "answer": "A"},
                         {"score": 1.0, "answer": "B"},
@@ -533,7 +533,7 @@ class TestAssessmentSerializer:
                 "error": "test error",
                 "min": None,
                 "max": None,
-                "semantic_id": "test-semantic-id",
+                "question_semantic_id": "test-semantic-id",
                 "answers": [
                     {"score": 2.0, "answer": "A"},
                     {"score": 1.0, "answer": "B"},
@@ -559,7 +559,7 @@ class TestAssessmentSerializer:
                 "error": "test error",
                 "min": None,
                 "max": None,
-                "semantic_id": "test-semantic-id",
+                "question_semantic_id": "test-semantic-id",
                 "answers": [
                     {"score": 2.0, "answer": "A"},
                     {"score": 1.0, "answer": "B"},
@@ -586,7 +586,7 @@ class TestAssessmentSerializer:
                 "error": "test error",
                 "min": None,
                 "max": None,
-                "semantic_id": None,
+                "question_semantic_id": None,
                 "answers": [
                     {"score": 2.0, "answer": "A"},
                     {"score": 1.0, "answer": "B"},
